Moves components to components/
